Add function license_list_overlap_non_hres()
This function tests if there is any overlap in non-hres license names
found in two lists. Hierarchal resource licenses are ignored in this
search. This is used by _future_run_test() when executed for a --test-only
job request. The --test-only job option does not support checking
hierarchal resource and this helps --test-only be more consistent with its
answers. Without this _future_run_test() can say a job overlaps with a hres
license when it actually doesn't. Additionally, even without this
_future_run_test() could also ignore running jobs that would overlap with
hres licenses. This just makes it always ignore hres licenses.

Ticket: 20564