sacctmgr - Fix ping persist_conn_t version
When opening persist_conn_t with slurmdbd, the version should be set to
the current protocol version, otherwise it will be set to
SLURM_MIN_PROTOCOL_VERSION by default. If an older sacctmgr is talking
to a newer slurmdbd, this will be an issue.

This commit ensures that the persist_conn_t is using sacctmgr's current
protocol version, which means that it can talk to a newer slurmdbd as
long as it's still a version that slurmdbd supports.

Ticket: 22793
Changelog: Fix sacctmgr ping to be able to connect to newer versioned
 slurmdbd.