Testsuite - Update test 37.5
Due to the changes in commit 06871cdd2f, the cluster that can run a job
soonest is the origin when --clusters are specified. The origin is still
the submit cluster when the --clusters option is not used.

Setting the viable siblings with --clusters means that the excluded
cluster will never be the origin. Instead, add a cluster feature and
specify the subset of clusters with --cluster-constraint so that the
job's origin is always the submit cluster as before.

Ticket: 11420