Fix x11 forwarding when using tls/s2n
In _net_forward(), we reuse the RPC connection to slurmstepd in a half
duplex eio objects. When we reuse the RPC, we need to make sure that it
doesn't get destroyed by tls_g_destroy_conn somewhere upstream, so we
set it to NULL.

This reset has happening prematurely before the tls_conn is handed over
to the half duplex eio objects. This commit ensures that the tls_conn is
reset after the handover.

Changelog: Fix x11 forwarding not working with TLS enabled.
Issue: 50464