Fix job update incorrectly rejected due to unintended tasks recompute
_figure_out_num_tasks() is called from two points (job creation and update).
Given the same job, it calculates a different number of tasks depending on
whether it's called in job creation context (job_ptr == NULL) or job update
context.

This causes problems because it's unconditionally called in _update_job()
even if the update RPC only changes unrelated fields (such as JobName, or
extending the TimeLimit) this function will recompute the ntasks and thus
the update will be rejected (because you can't change the number of tasks of
a running job). This leads to legal updates being rejected with incorrect
ESLURM_JOB_NOT_PENDING.

This patch gates the internal logic of _figure_out_num_tasks() behind a
check that we're actually requesting changes that could cause num_tasks to
be recomputed.

Ticket: 22720
Changelog: Fix cases of job updates incorrectly rejected when specifying
 modifications on fields unrelated to tasks computation (i.e. changing
 JobName).