Empty reservations cannot be created with ANY_NODES flag
When reservations had the ability to reserve TRES licenses and burst buffers
could be reserved through the TRES argument. The ANY_NODES flag should only
work with the burst buffer or licenses, but if the tres_str is set a bad
reservation can be created. Removing this and updating the error message
avoids the issue.

Changelog: Empty reservations can no longer be created with the ANY_NODES
 flag.
Ticket: 22852