tls/s2n - Change server config to request client certificate
This will allow mTLS to be used when clients have their own
certificate/key pairs loaded.

Note that this means all clients must connect with an s2n_config that
either has no certificate/key pair (no mTLS), or with an s2n_config that
has a certificate/key pair trusted by the loaded CA (mTLS).

If a client connects to a server with an s2n_config that has a
self-signed certificate loaded or is otherwise not trusted by the loaded
CA, the connection will be rejected as untrusted, i.e. even if the
client validates the server, the server will still reject the connection
because the client attempted to do mTLS.

Issue: 50465