tls/s2n - Ignore missing cert/key pair for client connections
Client connections will only load their own cert/key pair for mTLS. This
allows paths that don't need mTLS to still function.

Note that if mTLS is required on a connection for a certain operation,
it should be checked before doing that operation. This means that this
does not break security, i.e. a connection is not assumed to be mTLS
unless explicitly checked.

Issue: 50465