Prevent slurmd -C from potentially crashing
gres_get_autodetected_gpus() loops over every AutoDetect option, which
slurm may not be configured with. In the cases slurm is not configured with
the option the ext_lib_handle is not set in _get_gpu_type(). This is an
issue because gpu_plugin_fini() gets called and can call dlclose() on an
invalid ext_lib_handle from a previous iteration.

This prevents a potential crash of the slurmd -C by setting ext_lib_handle
to NULL after dlclose() is called on it.

Ticket: 22863
Changelog: Prevent slurmd -C from potentially crashing.