openapi/slurmctld - Don't dump all resv flags when none are set
The resv_desc_msg_t flag field needs to be NO_VAL64 if not set when it is
passed to the slurmctld. However, when this structure is being dumped if
the value is NO_VAL64 it outputs all flags. This is now prevented by
setting all NO_VAL64 flags to 0 before calling dump.

Changelog: openapi/slurmctld - Prevent dumping all reservation flags when
 none are set when posting reservations. This affects the following
 slurmrestd endpoints:
 'POST /slurm/v0.0.43/reservation'
 'POST /slurm/v0.0.43/reservations'