slurmcltd - Fix job_split_array decrement error
Caused by 0c46c43c42

Test 37.17 was failing due to an unregistered remote dependency on
the final array job. The symptom was that the last step of any array
with more that one task would stay pending forever.

Fixed a double decrement in job_array_split when task_cnt = 2.
Now the final REQUEST_SEND_DEP is sent to the remote cluster.

Ticket: 19658