Slurmstepd - Protect extern step from dying after signal delivery
slurmstepd includes a placeholder "sleep 100000000" process to enable
proctrack/linuxproc accounting.

When a signal is sent to the job, for example using the --signal option of
sbatch, the signal will be delivered to the extern step and to all its
processes, including the sleep process.

If the sleep process is signalled and terminates (most signals default
disposition is to terminate), the extern slurmstepd will also exit.

This can cause issues such as pam_slurm_adopt or X11 failing, and in worse
cases, the job itself may fail and be terminated since stepmgr runs as part
of the extern step.

This fix masks all signals for the "sleep 100000000" process except
SIGTERM, which is required to allow graceful termination.

Ticket: 22690
Changelog: Fix incorrect extern step termination when using sbatch or
 salloc --signal option.