Correctly wait for step terminate monitor thread
In situations where stepd takes long to start the monitor thread, we
could fall into a race that left the monitor thread waiting for a signal
that would never come.

This resulted in the job waiting in CG state for UnkillableStepTimeout
seconds and the node being incorrectly drained with "Kill task failed"
reason.

Ticket: 22864
Changelog: Fix a race that could incorrectly drain nodes due to "Kill
 task failed"