Remove bad sockets-per-node logic
This code existed from the beginning of cons_tres (see commit e82b9f17).
From the comments, the intention appears to be to remove sockets that do not
have enough GRES from consideration for the job. However, this is not doing
that. It is trying to remove sockets until we only have the number of
sockets to match --sockets-per-node, regardless of the resources on those
sockets. However, --sockets-per-node specified by the job is not actually
passed down to _build_sock_gres_by_topo(): the value in create_args->s_p_n
is determined from ntasks_per_socket (see _socks_per_node()) if set.
Usually, create_args->s_p_n is set to 1.

So, this code is instead removing sockets from consideration for the job that
otherwise would work for the job.

One result of this code is that the last sockets are allocated first:
socket 1 would be allocatd before socket 0 on a node.

For example:

   sbatch --gres-flags=enforce-binding -n1 --gpus=3 --wrap "sleep 60" -N1

Assuming a 2-socket node with 3 gpus per socket, this job is allocated on
socket 1 (the second socket) of the node with the corresponding gpus and
cores.

Another result is jobs that request --sockets-per-node > 1 can be rejected.
The removal of sockets fails to satisfy the --sockets-per-node requirement.
From sbatch(1):

    --sockets-per-node=<sockets>
    Restrict node selection to nodes with at least the specified number of
    sockets.

If you try to submit the previous job with --sockets-per-node=2, then this
job is rejected:

    sbatch --sockets-per-node=2 ...
    sbatch: error: Batch job submission failed: Requested node configuration
    is not available

This commit removes this bad logic from commit e82b9f17, and also removes
the field s_p_n from gres_sock_list_create_t since it is no longer used.