preempt/partition_prio - Do not default to PreemptMode PRIORITY
If not set, the partition's preempt_mode is set to NO_VAL16. The check for
if PREEMPT_MODE_PRIORITY is set did not verify before if it was NO_VAL16 or
not. This verifies that it's not before checking.

This fixes the case where a job with lower priority but on a high priority
tier partition, cannot preempt a job with higher priority but on a lower
priority tier partition. The priority should not matter if PreemptMode does
not include the PRIORITY flag.

Ticket: 22934
Changelog: preempt/partition_prio - Prevent partition PreemptMode
 defaulting to PRIORITY which caused jobs on higher priority tier partitions
 to not preempt jobs on lower priority tier partitions.