Fix incorrect gres validation on slurmd restart
On reconfiguration, GRES that are not binded to a full socket will be
set to INVAL state, making them unusable by the cluster.

This fixes an issue in which, if an INVALID node got their slurmd
restarted, gres config was not checked and INVALID state was removed.
This would ultimately leave a node with an incorrect configuration
usable by the cluster, causing errors on jobs allocated in that nodes.

Ticket: 22718
Changelog: Fix INVALID nodes going DRAIN after a slurmd restart without
 any gres.conf modification.