Testsuite - Fix test_144_2.py for enforce-binding
Test 144_2 is currently passing, but it's not testing enforce-binding.
The node configuration is unavailable because the test is requesting more
CPUs than exist. With the fix to EnforcePartLimits=NO,
test_gpu_socket_sharing_no_alloc started failing because the job is
accepted. Here, I increase the socket count so the limiting factor is
the gres socket binding.

Ticket: 22065
Cherry-picked: 7503579d6d