Fix slurmctld segfault when multiple QOS are specified in an array job
The prio_mult structure is set when a job has multiple partitions or QOSes.
If there was multiple QOSes sent to an array job, when the array job was to
be split the prio_mult would be ignored, leaving it to have the same
pointer as the job it was split from. When the job is deleted it results in
a double free which causes the slurmctld to crash. This will copy the
prio_mult if a QOS list exists.

Changelog: slurmctld - Prevent segfault when freeing job arrays that
 request one partition and a QOS list.
Ticket: 22455
Cherry-picked: f26944aca0