Fix s2n-tls stacktrace leak
Free any allocated s2n-tls stacktrace when cleaning up a TLS connection.

When stacktraces are enabled, s2n-tls allocates a stacktrace internally
when certain calls fail. Since the pointer to the stacktrace is kept in
thread-local storage, the memory has to be freed before the owning
thread exits. _cleanup_tls_conn() is a convenient place to do this.

Issue: 50439
Changelog: Fix memory leak when DebugFlags=TLS is configured