pmix - Fix return from function for PMIx v6.x
PMIx v6.0.0 calls to our _client_connected() function, but now it provides a
callback function. Until now the callback parameter was always NULL and it
was just ignored. Now we need to execute the callback if it exists, or
return PMIX_OPERATION_SUCCEEDED instead of PMIX_SUCCESS if it does not
exist. The user application will otherwise get stuck when using PMIx v6 with
Slurm.

This change is done in PMIx v6 to be compliant with the PMIx 4.1 standard
which requires to return PMIX_SUCCESS if the callback is going to be called,
and PMIX_OPERATION_SUCCEEDED if the request is immediately processed and the
callback will not be called.

This function is used to notify the host server that a client connected to
this server.

Ticket: 22956