slurmctld - Fix dependencies on remote array jobs
Fixes a regression in test37.17 introduced by 8f3392573c, which
added a check that array_task_ids are equal in _find_dependency.
Since _find_dependent_job_ptr sets array_task_id of NO_VAL to
INFINITE when an array job is found, the remote cluster would
respond with a different array_task_id than the one in the request.
This caused remote dependency updates to arrays to be ignored
if the dependency string was specified without "_*". For example,

sbatch --dependency=<job_id> <prog>

would never be fulfilled if <job_id> referred to a remote array job.

The local cluster has no way to know if a remote job id without
"_*" is an array. So the remote cluster responds with the
parsed array_task_id, not the resolved array_task_id in
dep_update_origin messages.

Ticket: 19658