Disallow new incoming RPCs for adding extern pids when terminating
When we are terminating the step we do not want any new incoming RPCs to
add pids to the extern step, as this operation can call the already
deinitialized plugins.

Additionally, the step state locking prevents a step from terminating
while an extern step is being added, ensuring that all extern pids are
tracked correctly to prevent any errors.

Ticket: 22770
Cherry-picked: 49e90498ac