Don't update avail_res_cnt
The avail_res_cnt should reflect only the current state of the node.
This count is utilized by the function common_topo_choose_nodes() to
eliminate nodes with the lowest available resources. The function
eval_nodes_cpus_to_use() adjusted this count to align with the proposed
allocation, but this modification is not relevant for
common_topo_choose_nodes().