slurm.spec - lower libyaml version requirement
Reduce the libyaml-devel version requirement in slurm.spec to 0.1.7 which
is used in RHEL 8. This package version requirement is not checked in
source builds, and is expected to work in all currently supported
platforms. The version check was set to that of RHEL 9 while adding the
"--with-yaml" flag in 78c8d4a32a. Since we are now testing on earlier
libyaml versions in CI it is safe to relax this version check.

Ticket: 22943
Changelog: Lower rpm libyaml version requirement to version in RHEL 8
Cherry-picked: d8e475093b