Testsuite - Update test37.10
The meaning of '-M' changed in commit 2287afd80da so that jobs are only
cancelled if the specified clusters include the job's origin cluster.

Rewrite test37.10 to check that jobs are cancelled when the origin
cluster is provided, but not when remote clusters are specified.
Also add a new sub test to verify that the --ctld flag cancels jobs.

Ticket: 11420