Docs - Remove reference to pacemaker in man 5 slurm.conf
Having an external entity manage and move the IP for controller is not
recommended due to the need for hostname and IP match which could be more
reliably handled via a container. Virtual IP (VIP) should not be
specified in a SlurmctldHost but instead in SlurmctldAddr.

Ticket: 22342
Cherry-picked: bdd95c356b