accounting_storage/mysql - Fix segv while gathering jobs with suspend time
While preparing to return a list of jobs that includes ones that have
been suspended and a truncate time condition is given (via "sacct -T" for
example), slurmdbd can crash if the database server goes down. Fix that.

Ticket: 23073
Changelog: Fix slurmdbd crash when preparing to return a list of jobs
 that include ones that have been suspended.