mpi/pmix - Lift symlink-no-follow restriction for trusted sources
Relax the restriction to the first open() of the parent directory and
subsequent fstatat() in pmxip_mkdir(), responsible for creating
{cli,lib}_tmpdir PMIx directories. The fstatat() only applies in
multiple-slurmd mode.

This restores the ability to configure symlinks and use PMIx for the now
considered trusted directories:

- slurm.conf SlurmdSpoolDir for lib_tmpdir
- mpi.conf PMIxCliTmpDirBase for cli_tmpdir
- slurm.conf TmpFS for cli_tmpdir

Which was restricted in Slurm 23.02.6 commit 88a473fd8f6.

The rest of sources will be considered untrusted and the symlink follow
restriction will remain applicable.

Changelog: mpi/pmix - Restore the ability to follow symlinks on PMIx cli/lib
 temporal directory creation when considered trusted (i.e. SlurmdSpoolDir,
 TmpFS, PMIxCliTmpDirBase). This ability was lost in Slurm 23.02.6.
Ticket: 17890