Fix the 'is-busy' check for a node shared between partitions
During scheduling, when a node belongs to multiple partitions it can be
marked as busy as though it were running jobs on other partitions when it is
not, in fact, doing so.

This is because, after a partition has run any job on such a node, it will
have a non-null value for row_bitmap for that node in some row in the
`select_part_record` linked list. However, it is the actual number of set
bits in these (per-node) core bitmaps that indicate whether that node is
running jobs for said partition (and on which cores).

 When _is_node_busy() is checking for different partitions running sharing
jobs on the same node, a job should not go unscheduled if one of those
partitions is not running jobs on that node and, equivalently, there are no
bits set in the row_bitmap for that node.

Ticket: 22228
Changelog: Fix the 'is-busy' check for a node shared between partitions