sbatch - Fix bad buffer truncation when retrieving user environment
In cases where --get-user-env (or any other --export flag that
implicitly uses it) is used for sbatch, if the installation path
leading to the slurmstepd binary is too long, it will fail the job
and requeue+hold it.

This happens because the internal mechanism used to retrieve the
environment uses a static buffer, which also needs to contain the
slurmstepd path. If the path gets too long, the buffer
ends up truncated.

This change removes the static buffer and replaces it with a
dynamically allocated one.

Ticket: 23096
Changelog: sbatch - Fix case where --get-user-env and some --export
 flags could make a job fail and get requeued+held if Slurm's
 installation path was too long.