Always set the qos_ptr to the one given instead of only on ACCT_POLICY_REM_SUBMIT
Originally we had this set things to leave on the highest priority qos_ptr to,
in theory, save time later so this would be the first one evaluated. But in the
coming commits this will change. We plan to have reservations gate on QOS. So
if the highest priority QOS is requested, but the reservation doesn't allow
that QOS it doesn't help to set it to that and could prevent the correct QOS
from being used.

Issue: 50466
Ticket: 23173