Testsuite - Improve test21.21 to avoid false failures
This is a continuation of the work done in ded15ae7d9, 961d0581c5 and
912bfb1f1e, for a very similar issue.

In this case the issue was that the jobs submitted were too short (10s)
and the scheduler was taking to much to schedule all the jobs and when
the grpjobsub was supposed to hit the first jobs were already done.
Now all job are submitted together as in the other commits and the waiting
for all jobs to be run is quicker.
And the jobs are longer, just in case.

Ticket: 21113