slurmctld - Make sure the child slurmctld relocks its pidfile
Before this patch, a reconfigure would cause the new slurmctld process
to not lock the pidfile. This meant that a new slumctld process could be
started and find the pidfile unlocked, and then trying to start without
killing the already running slurmctld. The newly started slurmctld would
modify the pidfile and likely fataling due to needed ports being in use by
the old slurmctld.

Ticket: 23221
Changelog: slurmctld - Relock the controller's pidfile on a reconfigure.