Remove comment originally added by Brian
There were a few issues with this comment. The one that caught my eye
was the duplicated function name (job_record_pack_details_common), then
I dug deeper and there are also other nested functions that should be
listed for completeness.

The amount of code churn that would result from bumping every outer
(or inner) block is excessive and the only benefit is keeping the protocol
version numbers in lockstep. I discussed the issue with him and the consensus
was that it is probably best to simply remove it.

Ticket: 23220