scontrol (show job) - add SubmitLine field
This patch series adds to scontrol the capability of displaying a job's
associated submit line. This is the full command line the user types when
submitting a job with sbatch/srun/salloc.

This information is also accessible via sacct but that requires a slurmdbd
whereas scontrol operates on information from the slurmctld only.

Ticket: 23220
Changelog: scontrol - Add "SubmitLine" field to show job subcommand's
 output.