Fix stepmgr enabled srun allocations failing when excluding nodes
This happens when a stepmgr enabled srun gets an allocation and excludes
nodes. The srun doesn't get allocated on the excluded nodes. The step
then tries to avoid the excluded nodes, but for stepmgr enabled steps
the stepmgr only has the allocated nodes in its node_record_table_ptr so
the stepmgr fails to find the requested excluded node and fails to
exclude the node -- that hasn't been allocated. The stepmgr shouldn't
even try to exclude a node that hasn't been allocated to it. So just
ignore the excluded nodes instead of throwing an error when stepmgr
encounters a nodename that is outside its allocation.

Changelog: Fix stepmgr enabled srun allocations failing when excluding
 nodes.
Ticket: 23311