Handle failure in update_job_dependency on a remote cluster
When the remote cluster received a dependency that caused
update_job_dependency() to fail, an error was logged and the dependency was
ignored. That resulted in the dependency on that job staying unfulfilled
forever.

This commit marks the dependency as failed and updates the job on the origin
cluster. This allows the dependency to be treated just like any other
dependency. If the dependency is AND'd with other dependencies, the job is
held with the reason DependencyNeverSatisfied. If the dependency is OR'd,
then only the invalid dependency is marked as failed and the job remains
pending waiting for results from other dependencies.

This behavior is different from a dependency on a local job id: if
update_job_dependency() fails on the origin cluster (or there is no
federation), then the job submission is rejected. Unfortunately, this is not
possible to do for a remote dependency, which must be tested on the remote
cluster after the job submission has been accepted. So, the best we can do
is to mark the dependency as failed.

This commit also documents this behavioral difference for afterok/afternotok
dependencies on local and remote jobs.

Ticket: 19658
Changelog: Mark afterok/afternotok dependencies on a non-existent remote job
 id as failed. Previously the dependency would remain unfulfilled
 indefinitely.