Update conmgr_threads description in slurm.conf man page
In the related ticket, the customer was confused about the ways in which
changing the conmgr_threads parameter would impact the speed of RPC
processing on their cluster. This commit uses more verbose language to
expand upon the impact that conmgr_threads has on connection processing
and performance.

Ticket: 23166
Cherry-picked: 9974f83833