Move error log for finding pid in task
The pid is used to try to find a task in a step. If a task is not found
(task == NULL), then SLURM_ERROR is returned.

Later, the same task is checked if it is NULL. This check is redundant,
and the error it would print would actually never be printed since the
function would have returned earlier.

This commit moves the error log so that it can actually be printed.

Ticket: 20458