Modify --wait-for-children to allow bad exit from parent
When using --wait-for-children and the parent process in a task exits
with a non-zero return code, the task is ended.

This commit modifies this behavior and allows a job run with
--wait-for-children to allow parent processes in a task to exit with a
non-zero return code without ending the task.

Ticket: 20458
Changelog: When using --wait-for-children, a task's behavior in regards
 to the parent process exit will now depend on --kill-on-bad-exit. If
 --kill-on-bad-exit=1 and the parent process exits non-zero, the task
 will end. If --kill-on-bad-exit=0 and the parent process exits with an
 error, the task will continue.  Note that default KillOnBadExit setting
 in slurm.conf is 0, which will result in different behavior for
 --wait-for-children as described above.
Cherry-picked: 5f191368bb