scontrol - Unify logic for printing multiple optional fields on a line
This commit unifies logic used for printing multiple optional fields on
the same line by adding the _add_to_line() function, which aims to
prevent bugs when modifying/adding new fields.

This new logic is applied to "Line 18" (BatchFeatures/BatchHost) and
"Line 38" (mostly GRES options).

This fixes a bug with the "GresAllowTaskSharing=Yes" field only printing
when something else on Line 38 was set. If "GresAllowTaskSharing=Yes"
was the only field to be printed, it would not print.

Changelog: scontrol - Fix GresAllowTaskSharing not always appearing in
 show jobs output.