conn - Return -1 for getting fd when conn doesn't exist
conn_g_get_fd is used widely in the codebase primarily for code paths
that were converted from using a raw fd to the new conn_t struct.

conn_g_get_fd asserts that the connection already exists and is
connected (in a non-TLS environment, this generally equates to a
successful connect() call), however there are code paths that were
designed (whether intentional or not) to operate on unopened fd's.

This commit changes the behavior conn_g_get_fd to conform to this fd
connection ambiguity, and allows calls for unopened connections (conn_t
is NULL).

Ticket: 23387
Cherry-picked: 2b0d100fc5