half_duplex - Add conn to remote_to_local_eio object
The remote_to_local_eio object is already tracking the underlying file
descriptor for the remote connection, but it also needs the conn plugin
interface connection.

This allows eio to peek on the connection to see if there's data ready
to be read on the connection. With the tls/s2n conn plugin interface
implementation, it's possible that the s2n-tls library will read all the
data on the underlying file descriptor before the data is actually
consumed via conn_g_recv.

With this commit, eio will properly track the remote connection of
half_duplex and not miss when data is ready to be read on it.

Ticket: 23026
Cherry-picked: 3b57acf6c7