eio - Use conn_g_peek before running poll
If the connection is using tls/s2n, then it's possible that on the
previous read, the s2n-tls library actually read all the data on the
underlying file descriptor, despite the data not actually being consumed
yet.

This means that even though data is ready to be read via conn_g_recv on
the connection, a poll() on the underlying file descriptor may show no
events and would block until the next event.

This commit ensures that conn_g_peek is called to check for any data
that is already ready to be read on the connection before running a
blocking poll() that may block indefinitely (at least until the next
event or timeout).

Ticket: 23026