half_duplex - Handle blocking write
The remote side of a half_duplex connection is always non-blocking. This
is necessary for the behavior of conn_g_recv. If conn_g_recv is called
for a blocking connection, it may wait until 'n' bytes are received on
the connection. This is not desirable, since half_duplex is designed to
read whatever data is available on the connection and immediately
forward it. This is why the connection needs to be non-blocking.

Due to the connection now being non-blocking, conn_g_send may return
with EWOULDBLOCK, which needs to be handled by the caller. This commit
handles EWOULDBLOCK, and ensures that all of the data that was read on
the local connection is written to the remote connection.

Ticket: 23026
Cherry-picked: f6d0b4f6a1