half_duplex - Handle blocking read
The remote side of a half_duplex connection non-blocking when TLS is
enabled. This is necessary for the behavior of conn_g_recv. If
conn_g_recv is called for a blocking connection, it may wait until 'n'
bytes are received on the connection. This is not desirable, since
half_duplex is designed to read whatever data is available on the
connection and immediately forward it. This is why the connection needs
to be non-blocking.

With the tls/s2n implementation of the conn plugin interface,
conn_g_recv may return with EWOULDBLOCK, despite poll() showing POLLIN
on the connection. This commit handles this case to ensure that the
connection is not incorrectly closed.

Ticket: 23026
Cherry-picked: 4947db4ad8