Keep duplicates on remote sibling
The remote sibling should not remove duplicates because the local
cluster expects an exact one-to-one correspondence between it's
depend_lists and the remote ones. In other words, removing duplicates
could result in a job that stays dependent forever.

Before a job with a dependency like '<job_id>,<job_id>_*', where the
job is an array on the remote cluster, remained pending forever. This
is because the local cluster couldn't resolve the id but the remote
knew that they are duplicates and only cleared '<job_id>_*' on completion.

Ticket: 19658
Changelog: Federated clusters update every registered remote dependency,
 including duplicates. This prevents jobs from pending forever due to some
 dependencies never being fulfilled.