Fix MaxJobCount off-by-one error
When job arrays were added, all incoming job requests would check how many
jobs were in the request against the MaxJobCount to ensure the number of
jobs could not overshoot the maximum. We compare if the total is >= to the
max, which gives us an off-by-one error.

Changelog: A cluster will start the MaxJobCount of jobs and not one less.
Cherry-picked: 4953797d2b