jobcomp - Fix slurmctld segfault if SIGUSR2 is received early
Conmgr registers the signal handlers before some plugins are initialized.
In this case the handler _on_sigusr2() called jobcomp_g_set_location(),
and if this happens immediately after the signal handler is registered
and before read_slurm_conf() is called, then the jobcomp plugin will be
uninitialized and it will segfault.

The jobcomp_g_set_location() was originally coupled with jobcomp_g_init()
but this changed after the introduction of conmgr and the commit a512ffc99a.

So we need to ensure jobcomp_g_init() is called before
jobcomp_g_set_location(), and in this case the less intrusive and
easiest way is the one in this commit. We don't expect that other calls
will be made before this call.

Changelog: Fix slurmctld segfault when SIGUSR2 is received early and
 jobcomp plugin is enabled.
Ticket: 23411
Cherry-picked: 7dd718c1fe