Add TLS certificate to het job requests
srun and other user commands require a self-signed ephemeral TLS
certificate in order to open servers which clients can connect to and
trust.

This certificate is meant to be sent to slurmctld upon job submission to
store in the job record for when it is needed. This happens in
slurm_allocate_resources_blocking for normal non-heterogeneous jobs, but
this functionality was missing from slurm_allocate_het_job_blocking.

This commit fixes heterogeneous jobs in a TLS enabled environment by
properly adding the already generated public certificate to each
heterogeneous job description during submission.

Ticket: 23273
Changelog: tls/s2n - Fix heterogeneous jobs failing to run in a TLS
 enabled environment.