Add new TRES_STR_FLAG_ALLOW_AMEND and TRES_STR_FLAG_COMB_AMEND
These new flags will be used to determine different modes of operation
when operating with TRES strings containing amending TRES.

In the context of these changes, an amending TRES is a TRES that is
constructed by parsing "+=" or "-=" syntax from sacctmgr when
specifying TRES alterations. The value for these TRES will be
internally codified as +<value> and -<value>.

For example:
  sacctmgr modify account rzarco set maxtrespj=cpu=5,mem+=2,node-=8

This will be treated as a string containing amending TRES, and
it will be internally translated to:

  cpu=5,mem=+2,node=-8

Further commits will allow this new format to be used for TRES
manipulations.

Issue: 50369