Prevent the forward message timeout from increasing down the chain
When the 24.05 protocol is used, tree_depth is not packed because it
didn't exist yet. In 24.11+, the timeout that each level waits is passed
in the header's timeout and if tree_depth was not set the recipient
would multiply the timeout by 2 * current depth, thus increasing the
timeout as the message went down the forwarding chain. This commit
prevents that for the slurmd's containing this commit.

Ticket: 23451
Changelog: Prevent healthy nodes being marked as unresponsive due to
 forwarding message timeouts increasing as the tree is traversed. The
 issue occurred if Slurm was running with a mix of 24.05- and 24.11+
 slurmds. This only fixes 25.05+ slurmds.