Fix use-after-free issue with the ident argument to openlog()
The syslog subsystem apparently stores this as a pointer, and not as an
independent copy. When log_set_prefix() is called in slurmscript or
slurmstepd we need to call closelog() to invalidate that pointer, and
ensure that the ident string will be updated correctly in successive
logging calls.

Regression from 669e20e297.

Changelog: slurmscriptd,slurmstepd - Fix use-after-free issue with the
 "ident" string when logging to syslog.
Ticket: 23586
Cherry-picked: 043884bdde