slurmctld - Prevent reserved gres from being used by the wrong jobs
Before it was possible for the gres in a reservation to be used by a job
not requesting the reservation. This was because reservations reserve a
specific gres bit, but sock_gres->bits_any_sock's bitmap still included all
the gres bits, even those allocated to reservations. Additionally, before
it was possible for a job requesting the reservation to be allocated a gres
bit that was not reserved due to the same reason.

This modifies _handle_gres_exc_bit_and_not(), now renamed to
_handle_gres_exc_bit_restrict(), so that now in the case of
resv_exc_ptr->gres_js_inc the bits_by_sock bitmap is restricted to only
gres bits in the reservation. This function is now also called on
sock_gres->bits_any_sock in the case no core topology is defined, which
can be the case for gres with a type defined.

Ticket: 21862
Changelog: slurmctld - Fix how gres with cores or a type defined are
 selected to prevent jobs not using reservations from being allocated
 reserved gres and vice versa.
Cherry-picked: 951427ba70