mpi/pmix - Remove incorrect logic to filter PMIxEnv variables
In ticket 9395 comment 29, PMIxEnv was originally supposed to only set env
vars that were prefixed with PMIX or UCX. However, the logic to due so
never worked as the xstrncmp should not be negated and the operator &&
acted incorrectly, so the logic allowed all environment variables to be
set. The docs for PMIxEnv also does not mention that restriction.

So since it is not documented and never worked this just removes the logic
which will allow for more flexibility in the future.

Ticket: 23080