Prevent hostlist_shift_iterators() from making hostlist iters repeat hr
Before when deleting a host range, hostlist_shift_iterators() would set the
i->idx to the host range before the deleted one. It also would not reset
the i->depth to -1. For example, if you have a hostlist "a[1-2],b1,c1", and
hostlist_remove() was called while the hostlist iterator was on b1, the
next call to hostlist_next() would return a2 when it should be c1.

This make it so if the host range i->idx was referencing was deleted then
i->idx will be set to the new index referencing the host range after the
one that was deleted, and it will reset the depth.

This issue could cause memory corruption in _get_alias_addrs() when writing
to forward->alias_addrs.node_addrs[addr_index] because the hostlist
iterator loop could repeat host ranges leading to addr_index going out of
bounds.

Ticket: 23676
Co-authored-by: Megan Dahl <megan@schedmd.com>
Changelog: Prevent potential memory corruption while forwarding messages
 that require addresses to be packed.