Fix CgroupPlugin=Autodetect when built without cgroup support
Even if Slurm is built without cgroup it is still possible to configure
cgroup.conf with CgroupPlugin=autodetect. In that case if the system
does have a cgroup filesystem, autodetect will return a version, which
will then fail to load because the cgroup will not be detected.

To workaround this just return "disabled" in CgroupPlugin.

Changelog: CgroupPlugin=Autodetect will always return "disabled" when
 Slurm is built without cgroup support.